refactor for fc and iscsi generic code
